+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 
 
 ## Unreleased
-
+- Added: Optional parameters "security_allow_no_token"
+- Added: Optional parameters "security_token_lifetime_no_refresh"
 
 ## [0.1.17] 2020-12-18
 - Added: optional SMS cleanup lambda